requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1

# ML / Vision
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.40.0
sentence-transformers>=2.7.0
Pillow>=10.3.0
opencv-python-headless>=4.9.0

# Privacy preprocessing
# face detection uses OpenCV DNN or mediapipe
mediapipe>=0.10.14

# Storage / Data
sqlalchemy==2.0.30
aiosqlite==0.20.0
greenlet>=3.0.0
pandas>=2.2.0
pyarrow>=16.0.0

# Embeddings / Search
faiss-cpu>=1.8.0

# Dashboard
streamlit>=1.35.0
plotly>=5.22.0
altair>=5.3.0

# Image utils
piexif>=1.1.3

# Testing
pytest>=8.2.0
pytest-asyncio>=0.23.0
anyio[asyncio]>=4.3.0
httpx>=0.27.0

# Dev
python-dotenv>=1.0.1
rich>=13.7.0
tqdm>=4.66.0
